# Conflicts:
#	src/Oro/Bundle/SearchBundle/Resources/config/services.yml
#	src/Oro/Bundle/WorkflowBundle/Resources/config/services.yml
#	src/Oro/Bundle/WorkflowBundle/Resources/public/js/transition-executor.js

+++ src/Oro/Bundle/AttachmentBundle/Manager/AttachmentManager.php
@@ -4,6 +4,8 @@ namespace Oro\Bundle\AttachmentBundle\Manager;
 
 use Doctrine\ORM\EntityManager;
 
+use Gaufrette\Stream;
+
 use Symfony\Component\Security\Core\Util\ClassUtils;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Bundle\FrameworkBundle\Routing\Router;
@@ -28,6 +30,8 @@ class AttachmentManager
     const READ_COUNT = 100000;
     const DEFAULT_IMAGE_WIDTH = 100;
     const DEFAULT_IMAGE_HEIGHT = 100;
+    const SMALL_IMAGE_WIDTH = 32;
+    const SMALL_IMAGE_HEIGHT = 32;
 
     /** @var Filesystem */
     protected $filesystem;
@@ -105,7 +109,7 @@ class AttachmentManager
     public function preUpload(File $entity)
     {
         if ($entity->isEmptyFile()) {
-            if ($this->filesystem->has($entity->getFilename())) {
+            if ($entity->getFilename() !== null && $this->filesystem->has($entity->getFilename())) {
                 $this->filesystem->delete($entity->getFilename());
             }
             $entity->setFilename(null);
@@ -131,7 +135,7 @@ class AttachmentManager
                 $entity->setFileSize($file->getSize());
             }
 
-            $entity->setFilename(uniqid() . '.' . $entity->getExtension());
+            $entity->setFilename($this->generateFileName($entity->getExtension()));
 
             $fsAdapter = $this->filesystem->getAdapter();
             if ($fsAdapter instanceof MetadataSupporter) {
@@ -164,17 +168,8 @@ class AttachmentManager
      */
     public function copyLocalFileToStorage($localFilePath, $destinationFileName)
     {
-        $src = new LocalStream($localFilePath);
-        $dst = $this->filesystem->createStream($destinationFileName);
-
-        $src->open(new StreamMode('rb+'));
-        $dst->open(new StreamMode('wb+'));
-
-        while (!$src->eof()) {
-            $dst->write($src->read(self::READ_COUNT));
-        }
-        $dst->close();
-        $src->close();
+        $srcStream = new LocalStream($localFilePath);
+        $this->copyStreamToStorage($srcStream, $destinationFileName);
     }
 
     /**
@@ -441,5 +436,56 @@ class AttachmentManager
             $this->associationManager->getSingleOwnerFilter('attachment'),
             RelationType::MANY_TO_ONE
         );
+    }
+
+    /**
+     * Copy attachment file object
+     *
+     * @param File $file
+     *
+     * @return File
+     */
+    public function copyAttachmentFile(File $file)
+    {
+        $fileCopy = clone $file;
+        $fileCopy->setFilename($this->generateFileName($file->getExtension()));
+
+        $sourceStream =  $this->filesystem->createStream($file->getFilename());
+        $this->copyStreamToStorage($sourceStream, $fileCopy->getFilename());
+
+        return $fileCopy;
+    }
+
+    /**
+     * Copy stream to storage
+     *
+     * @param Stream $srcStream
+     * @param string $destinationFileName
+     */
+    protected function copyStreamToStorage(Stream $srcStream, $destinationFileName)
+    {
+        $dstStream = $this->filesystem->createStream($destinationFileName);
+
+        $srcStream->open(new StreamMode('rb+'));
+        $dstStream->open(new StreamMode('wb+'));
+
+        while (!$srcStream->eof()) {
+            $dstStream->write($srcStream->read(self::READ_COUNT));
+        }
+
+        $dstStream->close();
+        $srcStream->close();
+    }
+
+    /**
+     * Generate unique file name with specific extension
+     *
+     * @param string $extension
+     *
+     * @return string
+     */
+    protected function generateFileName($extension)
+    {
+        return sprintf('%s.%s', uniqid(), $extension);
     }
 }
